Interp

Description

Implementation an interpreter for RFun

The design is intended to follow closely* the design of the RFun paper:

T. Yokoyama, H. B. Axelsen, and R. Gluck Towards a reversible functional language LNCS vol. 7165, pp. 14--29, 2012

I know that there are an obvious reader monad below.

Interpreter main

type EvalTrace = [String]

Evaluating with return either a result of an ErrorRFun

type ErrorRFun = String

An error is a String

newtype Eval a

Constructors

E

Fields

runE :: WriterT EvalTrace (Either ErrorRFun) a

runEval :: Eval a -> Either ErrorRFun (a, EvalTrace)

evalMaybe :: ErrorRFun -> Maybe a -> Eval a

Making an Maybe an Eval

failEval :: ErrorRFun -> Eval a

Simple fail

runProg :: Ident -> Value -> FuncEnv -> Either ErrorRFun (Value, EvalTrace)

Interpreting an rFun program

Substitutions and functions on these

type Substitution = Map Ident Value

A substitution is a mapping from a Ident (string) to a value

idSub :: Substitution

Make a empty substitution

newSub :: Ident -> Value -> Substitution

Make a substitution of one variable

peakValue :: Ident -> Substitution -> Maybe Value

Lookup a value in a substitution without any limitations -- used in function lookup

lookupValue :: Ident -> Substitution -> Eval Value

Lookup a value in a substitution

divide :: [Ident] -> Substitution -> Eval (Substitution, Substitution)

Divides a substitution into two. First part contains the mappings contained in the list of Idents, the second the rest.

lookupDivide :: Ident -> Substitution -> Eval (Value, Substitution)

Lookup a value in a substitution. Returns the value of the identifier and the substitution with the identifier removed.

disUnion :: Substitution -> Substitution -> Eval Substitution

Finds the disjoint union between two substitutions

disjointUnion_M :: Eval Substitution -> Eval Substitution -> Eval Substitution

Finds the union between two disjoint substitutions (no overlap in idents).

disjointUnions_M :: [Eval Substitution] -> Eval Substitution

Finds the union between a list of disjoint substitutions.

Program functions

lookupFunction :: FuncEnv -> Ident -> Eval (LExpr, Expr)

Lookup a function in the function environment

The interpreter

evalDupEq :: Value -> Eval Value

Eq/Dup operator (Eqs. 3 and 4, p. 17)

evalRMatchS :: Value -> LExpr -> Eval Substitution

R-Match (Fig. 2, p. 18) that returns a substitution. Returns a substitution

evalRMatchV :: Substitution -> LExpr -> Eval Value

R-Match (Fig. 2, p. 18) that returns a value

evalFunS :: FuncEnv -> Substitution -> Ident -> LExpr -> Value -> Eval Substitution

Function calls: Fig 3, p. 19, FunExp Function calls in a sub part of lets

evalFunV :: FuncEnv -> Substitution -> Substitution -> Ident -> LExpr -> Eval Value

Function calls that returns a value

evalExpS :: FuncEnv -> Expr -> Value -> Eval Substitution

Expressions: Fig 3, p. 19 (not FunExp) that returns substitution

evalExpV :: FuncEnv -> Substitution -> Expr -> Eval Value

Expressions: Fig 3, p. 19 (not FunExp) that returns value

checkLeaves :: (Value -> LExpr -> Eval c) -> Value -> [LExpr] -> Maybe Value

This function is helper for the caseOf

checkLExprs :: (Value -> LExpr -> Eval c) -> Value -> Substitution -> [LExpr] -> Maybe Substitution

This function is helper for the caseOf

findSubIndex :: (a -> Eval b) -> [(Int, a)] -> Maybe (Int, b)

Finds the minimum index of a case-leave to which a eval-function matches. The list is indexed from 0; different from the paper!!!!

leaves :: Expr -> [LExpr]

As defined in Footnote 1, p 19.

findVars :: LExpr -> [Ident]

Finds the list of all variables in a left expression