Preparse

Description

Preparse for rFun language

Current purpose of this is to generate the function environment and de-sugar the function level pattern matching to case expressions.

Pre-parse / de-sugar call function

runPreparse :: Program -> FuncEnv

Pre-parse / de-sugar call function

applyToFunctionBody :: (Expr -> Expr) -> Program -> Program

General function that applies a function to all function bodies in a program.

Collecting functions of identical name for the function environment

functionList :: Program -> [Ident]

Generates a list of unique function names

programToFuncEnvSS :: Program -> [(Ident, [Func])]

Generates function environment from a list of functions

De-sugar pattern matching in arguments, while preserving order

desugarArgPatMatch :: [(Ident, [Func])] -> [(Ident, Func)]

De-sugar pattern matching in arguments, while preserving order

De-sugar function calls in expressions

De-sugar many let assignments